Fix arithmetic overflows in planning & experimental modules

+++ crates/logos-core/src/experimental/goal_fund_projector.rs
@@ -76,11 +76,13 @@ impl GoalFundProjector {
         for month_index in 1..=months {
             let mut allocated_this_month = 0_i64;
 
-            let month_start_days = (month_index - 1) * 30;
-            let month_end_days = month_index * 30;
+            let month_start_days = (u32::from(month_index) - 1) * 30;
+            let month_end_days = u32::from(month_index) * 30;
 
             for vest in &self.upcoming_vests {
-                if vest.days_to_vest > month_start_days && vest.days_to_vest <= month_end_days {
+                if u32::from(vest.days_to_vest) > month_start_days
+                    && u32::from(vest.days_to_vest) <= month_end_days
+                {
                     let safe_value = forecast_value_cents(
                         vest.avg_close_price_cents,
                         vest.units,
